Cleanup poco/Foundation: 'size_t' related warnings
 * eliminate conversion warnings for `size_t` related expressions.
 * convert a few `int` to `size_t`
 * SharedMemory files `_size` is now LONGLONG and mappings are correctly
	created for files larger than 65kb.

+++ Externals/poco/Foundation/include/Poco/SharedMemory_WIN32.h
@@ -98,7 +98,7 @@ private:
 	std::string _name;
 	HANDLE  _memHandle;
 	HANDLE  _fileHandle;
-	DWORD   _size;
+	LONGLONG  _size;
 	DWORD   _mode;
 	char*   _address;
 };
